pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "hemostat"
version = "0.1.0"
description = "Autonomous Container Health Monitoring System - Multi-agent system for Docker container health monitoring and remediation"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [{ name = "Team 1", email = "team@hemostat.local" }]
keywords = [
    "docker",
    "monitoring",
    "health-check",
    "multi-agent",
    "redis",
    "ai",
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: System :: Monitoring",
    "Topic :: System :: Distributed Computing",
]

# Core dependencies for Phase 1
dependencies = [
    "redis==5.0.1",              # Redis Python client with pub/sub support
    "python-dotenv==1.0.0",      # Environment variable loading from .env files
    "python-json-logger==2.0.7", # Structured JSON logging
]

[project.optional-dependencies]
# Phase 2 - Agent Implementations
agents = [
    "docker==7.0.0",              # Monitor/Responder agents for container management
    "langchain==0.1.0",           # LLM orchestration for Analyzer agent
    "langchain-openai==0.0.5",    # OpenAI provider integration for LangChain
    "langchain-anthropic==0.1.0", # Anthropic provider integration for LangChain
    "openai==1.10.0",             # OpenAI GPT-4 integration
    "anthropic>=0.16.0",          # Anthropic Claude integration (compatible with langchain-anthropic)
    "requests==2.31.0",           # HTTP client for Alert agent (Slack webhooks)
]

# Phase 3 - Dashboard
dashboard = [
    "streamlit==1.29.0", # Web UI framework for live monitoring dashboard
]

# Phase 4 - Testing & Monitoring
dev = [
    "pytest==7.4.0",          # Unit testing framework
    "pytest-asyncio==0.21.0", # Async test support
    "pytest-cov==4.1.0",      # Code coverage reporting
    "ruff==0.5.3",            # Fast Python linter and formatter
    "ty==0.0.1a25",           # Fast type checker from Astral (written in Rust)
    "pre-commit==3.6.0",      # Git hooks for code quality
]

# All optional dependencies
all = [
    "docker==7.0.0",
    "langchain==0.1.0",
    "langchain-openai==0.0.5",
    "langchain-anthropic==0.1.0",
    "openai==1.10.0",
    "anthropic>=0.16.0",
    "requests==2.31.0",
    "streamlit==1.29.0",
    "pytest==7.4.0",
    "pytest-asyncio==0.21.0",
    "pytest-cov==4.1.0",
    "ruff==0.5.3",
    "ty==0.0.1a25",
    "pre-commit==3.6.0",
]

[project.urls]
Homepage = "https://github.com/jondmarien/HemoStat"
Documentation = "https://github.com/jondmarien/HemoStat/wiki"
Repository = "https://github.com/jondmarien/HemoStat.git"
Issues = "https://github.com/jondmarien/HemoStat/issues"

[tool.uv]
# UV configuration for fast Python package management
managed = true
dev-dependencies = [
    "pytest==7.4.0",
    "pytest-asyncio==0.21.0",
    "pytest-cov==4.1.0",
    "ruff==0.5.3",
    "pre-commit==3.6.0",
    "ty==0.0.1a25",
]

[tool.hatch.build.targets.wheel]
packages = ["agents"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
addopts = "--cov=agents --cov-report=html --cov-report=term-missing"

[tool.coverage.run]
source = ["agents"]
omit = ["*/tests/*", "*/test_*.py"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
]

[tool.typos.default]
check-filenames = true
check-files = true

[tool.typos.default.extend-ignore-identifiers]
# Identifiers that may contain false positive typos
hemostat = "hemostat"
pubsub = "pubsub"
redis = "redis"
remediaton = "remediaton" # Custom spelling for internal use

[tool.typos.default.extend-words]
# Words to recognize as correct
aerosol = "aerosol"

[tool.typos.files]
ignore-files = [".git", ".venv", "htmlcov", "dist", "build", "*.egg-info"]

[tool.ruff]
# Fast Python linter configuration
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # Pyflakes
    "I",   # isort - import sorting
    "N",   # pep8-naming
    "UP",  # pyupgrade
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "ARG", # flake8-unused-arguments
    "SIM", # flake8-simplify
    "RUF", # Ruff-specific rules
]
ignore = [
    "E501",   # Line too long (handled by formatter)
    "ARG001", # Unused function argument (can be intentional)
    "ARG002", # Unused method argument (can be intentional)
]

[tool.ruff.lint.isort]
known-first-party = ["agents"]

[tool.ruff.lint.flake8-quotes]
inline-quotes = "double"

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
line-ending = "auto"

[tool.ty]
# Fast type checker configuration (from Astral, written in Rust)
# ty automatically discovers packages in the active virtual environment
